AMP-16697 SSC activity: edit and then change languate redirects to activtiy preview
git-svn-id: file:///svn/branches/AMP_2_8_RELEASE@22336 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/dgfoundation/amp/onepager/util/ActivityGatekeeper.java
@@ -72,7 +72,7 @@ public class ActivityGatekeeper {
             }
             else{
                 if ((keycode.get(id) == null && timestamp.get(id) == null && userEditing.get(id) == null) ||
-                    (keycode.get(id) != null && userEditing.get(id) == userId))
+                    (keycode.get(id) != null && userEditing.get(id).longValue() == userId.longValue()))
                     return REFRESH_LOCK_EXPIRED; //lock was cleared on purpose, page probably getting refreshed now
                 else
                     return REFRESH_LOCK_LOCKED;
